[DONT MERGE] demonstrate a simplified search
This demonstrates a backtracking search extending the query to the
right. It uses a bidirectional index and cursor.
Usage can be seen in "test/unit/search/search_ng_test.cpp".
Implementation can be found in "include/seqan3/search/search_ng.hpp"

This implements a new cursor called "bi_fm_index_cursor_ng"
which is a simplified version of the "bi_fm_index_cursor" and hopefully
easier to use.
For this to work "bi_fm_index.hpp" and "fm_index.hpp" had to be touched
(needed a friend declaration)